# apps/local-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexura Local Server API",
    "version": "0.0.1",
    "description": "The per-branch local server's real API surface -- LAN-only, offline-first (Auth/Distribution doc Part 2: this server has no public URL; it's reachable only on the hotel's own network). Every path below was extracted directly from the real Express route definitions in `src/routes/*.ts` (see `scripts/generate-openapi.mjs`), not hand-typed -- run `npm run openapi:generate` after changing routes to keep this in sync. Auth is a JWT in an httpOnly `access_token` cookie (12h, with an 8h offline-continuation grace period -- see `src/auth/tokens.ts`), not a bearer token in the traditional OpenAPI sense; represented here as a cookie API key for tooling compatibility.",
    "license": {
      "name": "Proprietary — internal use only"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local dev server"
    },
    {
      "url": "http://nexura.local:4000",
      "description": "Real branch deployment (Auth doc Part 2, Layer 3) -- LAN hostname, no public DNS/TLS. `localhost` above is genuinely correct for dev, not a placeholder; this second entry is the real production shape."
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  },
...
  "paths": {
...
    "/admin/users": {
      "get": {
        "operationId": "getAdminUsers",
        "tags": [
          "IT Admin"
        ],
        "summary": "User Management list (IT-01)",
        "description": "**Requires permission:** `admin:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      },
      "post": {
        "operationId": "postAdminUsers",
        "tags": [
          "IT Admin"
        ],
        "summary": "Invite a new user (IT-01)",
        "description": "Returns a one-time temp password. `role` validated against the live `roles` table.\n\n**Requires permission:** `admin:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "firstName",
                  "lastName",
                  "role"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/announcements": {
      "get": {
        "operationId": "getAnnouncements",
        "tags": [
          "Announcements"
        ],
        "summary": "Announcement list (CO-03)",
        "description": "`?archived=true` includes archived ones. Read counts are real, tracked via `announcement_reads`.\n\n**Requires authentication** (any role).",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          }
        }
      },
      "post": {
        "operationId": "postAnnouncements",
        "tags": [
          "Announcements"
        ],
        "summary": "Post an announcement",
        "description": "**Requires permission:** `announcements:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "body"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "body": {
                    "type": "string"
                  },
                  "targetAudience": {
                    "type": "string",
                    "default": "all"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/availability": {
      "get": {
        "operationId": "getAvailability",
        "tags": [
          "availability"
        ],
        "summary": "GET /availability",
        "description": "**Requires permission:** `rates:read` or `inventory:calendar` or `reservations:create`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/deposits": {
      "get": {
        "operationId": "getDeposits",
        "tags": [
          "deposits"
        ],
        "summary": "GET /deposits",
        "description": "**Requires permission:** `deposits:manage` or `finance:read`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      },
      "post": {
        "operationId": "postDeposits",
        "tags": [
          "deposits"
        ],
        "summary": "POST /deposits",
        "description": "**Requires permission:** `deposits:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/finance/folios": {
      "get": {
        "operationId": "getFinanceFolios",
        "tags": [
          "Finance & Billing"
        ],
        "summary": "Folio Management list (FI-01)",
        "description": "`?status=`, `?minBalance=true`, `?dateFrom=`/`?dateTo=` filters.\n\n**Requires permission:** `folio:read`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/guests": {
      "get": {
        "operationId": "getGuests",
        "tags": [
          "Guests"
        ],
        "summary": "Search guests",
        "description": "`?search=` matches first name, last name, or phone. No `search` param returns the branch's most recent 20 guests.\n\n**Requires authentication** (any role).",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          }
        }
      },
      "post": {
        "operationId": "postGuests",
        "tags": [
          "Guests"
        ],
        "summary": "Create a guest profile",
        "description": "FD-04 inline guest creation.\n\n**Requires permission:** `guests:create`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName"
                ],
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "idType": {
                    "type": "string"
                  },
                  "idNumber": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/inventory-calendar": {
      "get": {
        "operationId": "getInventoryCalendar",
        "tags": [
          "inventoryCalendar"
        ],
        "summary": "GET /inventory-calendar",
        "description": "**Requires permission:** `inventory:calendar` or `rates:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      },
      "patch": {
        "operationId": "patchInventoryCalendar",
        "tags": [
          "inventoryCalendar"
        ],
        "summary": "PATCH /inventory-calendar",
        "description": "**Requires permission:** `rates:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/inventory/dashboard": {
      "get": {
        "operationId": "getInventoryDashboard",
        "tags": [
          "Inventory"
        ],
        "summary": "Stock Dashboard (IV-01)",
        "description": "Low-stock count, total inventory value, and the 10 most recent transactions.\n\n**Requires permission:** `inventory:read`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/invoices": {
      "get": {
        "operationId": "getInvoices",
        "tags": [
          "invoices"
        ],
        "summary": "GET /invoices",
        "description": "**Requires permission:** `folio:read` or `finance:read`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      },
      "post": {
        "operationId": "postInvoices",
        "tags": [
          "invoices"
        ],
        "summary": "POST /invoices",
        "description": "**Requires permission:** `invoices:issue`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/lost-found": {
      "get": {
        "operationId": "getLostFound",
        "tags": [
          "Lost & Found"
        ],
        "summary": "Lost & Found log (HK-05)",
        "description": "**Requires authentication** (any role).",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          }
        }
      },
      "post": {
        "operationId": "postLostFound",
        "tags": [
          "Lost & Found"
        ],
        "summary": "Log a found item",
        "description": "**Requires permission:** `frontoffice:lostfound`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "description"
                ],
                "properties": {
                  "description": {
                    "type": "string"
                  },
                  "locationFound": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/rate-calendar": {
      "get": {
        "operationId": "getRateCalendar",
        "tags": [
          "rateCalendar"
        ],
        "summary": "GET /rate-calendar",
        "description": "**Requires permission:** `rates:read` or `rates:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/rate-plans": {
      "get": {
        "operationId": "getRatePlans",
        "tags": [
          "ratePlans"
        ],
        "summary": "GET /rate-plans",
        "description": "**Requires permission:** `rates:read` or `rates:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      },
      "post": {
        "operationId": "postRatePlans",
        "tags": [
          "ratePlans"
        ],
        "summary": "POST /rate-plans",
        "description": "**Requires permission:** `rates:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
    "/receipts": {
      "get": {
        "operationId": "getReceipts",
        "tags": [
          "receipts"
        ],
        "summary": "GET /receipts",
        "description": "**Requires permission:** `folio:read` or `finance:read`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      },
      "post": {
        "operationId": "postReceipts",
        "tags": [
          "receipts"
        ],
        "summary": "POST /receipts",
        "description": "**Requires permission:** `invoices:issue`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
  }
}